refactor: 여러 기록 인증 방법을 하나의 컴포넌트로 분리(이미지 제외)
