+++ components/molecules/record-mark/index.ts
@@ -1,0 +1,1 @@
+export * from './record-mark';
