+++ features/main/calendar/atoms/calendar-item-content.tsx
@@ -1,13 +1,10 @@
 import { useAtomValue } from 'jotai';
-import { useEffect, useRef, useState } from 'react';
 
 import { Image } from '@/components/atoms';
-import { SwimmerIcon } from '@/components/atoms/icons/swimmer-icon';
-import { Waves } from '@/components/atoms/waves';
+import { RecordMark } from '@/components/molecules/record-mark';
 import { calendarViewImageAtom } from '@/store';
 import { css, cx } from '@/styled-system/css';
 import { flex } from '@/styled-system/patterns';
-import { createGradient, getSwimColor } from '@/utils/visualization';
 
 import { StrokeInfo } from '../../time-line';
 import { MemoryType } from '../types';
@@ -20,9 +17,6 @@ interface ItemContentProps {
   imageUrl?: string;
 }
 
-// TODO: 로그인 이후 저장된 유저의 목표 거리로 수정 필요
-const goal = 1000;
-
 export const ItemContent = ({
   type,
   totalDistance,
@@ -30,31 +24,7 @@ export const ItemContent = ({
   isAchieved,
   imageUrl,
 }: ItemContentProps) => {
-  const ref = useRef<HTMLDivElement | null>(null);
   const isViewImage = useAtomValue(calendarViewImageAtom);
-  const [contentSize, setContentSize] = useState<{
-    width: number;
-    height: number;
-  }>({
-    width: 0,
-    height: 0,
-  });
-
-  const waves: Array<{ color: string; waveHeight: number }> = [];
-  if (strokes) {
-    strokes.forEach(({ name, meter }) => {
-      const color = getSwimColor(name);
-      waves.push({ color, waveHeight: meter / goal });
-    });
-  }
-
-  useEffect(() => {
-    if (ref.current)
-      setContentSize({
-        width: ref.current.offsetWidth,
-        height: ref.current.offsetHeight,
-      });
-  }, [isViewImage]);
 
   if (isViewImage && imageUrl)
     return (
@@ -72,59 +42,13 @@ export const ItemContent = ({
       </div>
     );
 
-  if (type === 'NORMAL')
-    return (
-      <div className={wrapperStyles}>
-        <div className={cx(layoutStyles, normalStyles)}>
-          <SwimmerIcon />
-        </div>
-      </div>
-    );
-  else if (type === 'SINGLE')
-    return (
-      <div className={wrapperStyles}>
-        {isAchieved ? (
-          <div className={cx(layoutStyles, singleAchievedStyles)} />
-        ) : (
-          <div ref={ref} className={layoutStyles}>
-            {ref.current && totalDistance && (
-              <Waves
-                width={contentSize.width}
-                height={contentSize.height}
-                waves={[
-                  {
-                    color: '#3b82f6',
-                    waveHeight: totalDistance / goal,
-                  },
-                ]}
-              />
-            )}
-          </div>
-        )}
-      </div>
-    );
-
-  const gradientProps = createGradient(waves.map(({ color }) => color));
-
   return (
-    <div className={wrapperStyles}>
-      {isAchieved ? (
-        <div
-          style={{ backgroundImage: `linear-gradient(0deg, ${gradientProps})` }}
-          className={layoutStyles}
-        />
-      ) : (
-        <div ref={ref} className={layoutStyles}>
-          {ref.current && totalDistance && (
-            <Waves
-              width={contentSize.width}
-              height={contentSize.height}
-              waves={waves}
-            />
-          )}
-        </div>
-      )}
-    </div>
+    <RecordMark
+      type={type}
+      totalDistance={totalDistance}
+      isAchieved={isAchieved}
+      strokes={strokes}
+    />
   );
 };
 
@@ -142,9 +66,4 @@ const layoutStyles = flex({
   justifyContent: 'center',
   rounded: '2px',
   overflow: 'hidden',
-});
-
-const normalStyles = css({ backgroundColor: 'blue.90' });
-const singleAchievedStyles = css({
-  backgroundColor: 'primary.swim.총거리.default',
 });
